+++ src/ansible_collections/solace/pubsub_plus/plugins/modules/solace_client_profile.py
@@ -30,7 +30,7 @@ options:
 extends_documentation_fragment:
 - solace.pubsub_plus.solace.broker
 - solace.pubsub_plus.solace.vpn
-- solace.pubsub_plus.solace.settings
+- solace.pubsub_plus.solace.sempv2_settings
 - solace.pubsub_plus.solace.state
 - solace.pubsub_plus.solace.broker_config_solace_cloud
 seealso: