+++ src/ansible_collections/solace/pubsub_plus/plugins/modules/solace_client_username.py
@@ -28,7 +28,7 @@ options:
 extends_documentation_fragment:
 - solace.pubsub_plus.solace.broker
 - solace.pubsub_plus.solace.vpn
-- solace.pubsub_plus.solace.settings
+- solace.pubsub_plus.solace.sempv2_settings
 - solace.pubsub_plus.solace.state
 author:
   - Ricardo Gomez-Ulmke (@rjgu)